// frontend/src/api/generated/model/protocolSummaryViewModel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { Crop } from './crop';
import type { ProtocolStatus } from './protocolStatus';

export interface ProtocolSummaryViewModel {
  id: string;
  code: string;
  version: number;
  crop: Crop;
  name: string;
  status: ProtocolStatus;
  targetCount: number;
}